requirements.txt
groq>=0.4.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
pypdf>=3.17.0
faiss-cpu>=1.7.4
streamlit>=1.22.0
